# Apply sanity caps on outlier values in performance reports