--- a/plugins/PagePerformance/Columns/TimeDomProcessing.php
+++ b/plugins/PagePerformance/Columns/TimeDomProcessing.php
@@ -30,7 +30,7 @@ class TimeDomProcessing extends Base
 
     public function configureMetrics(MetricsList $metricsList, DimensionMetricFactory $dimensionMetricFactory)
     {
-        $metric1 = $dimensionMetricFactory->createMetric(ArchivedMetric::AGGREGATION_SUM);
+        $metric1 = $dimensionMetricFactory->createMetric('sum(' . $this->getSqlCappedValue() . ')');
         $metric1->setName('sum_time_dom_processing');
         $metricsList->addMetric($metric1);