--- a/plugins/PagePerformance/Columns/TimeServer.php
+++ b/plugins/PagePerformance/Columns/TimeServer.php
@@ -30,7 +30,7 @@ class TimeServer extends Base
 
     public function configureMetrics(MetricsList $metricsList, DimensionMetricFactory $dimensionMetricFactory)
     {
-        $metric1 = $dimensionMetricFactory->createMetric(ArchivedMetric::AGGREGATION_SUM);
+        $metric1 = $dimensionMetricFactory->createMetric('sum(' . $this->getSqlCappedValue() . ')');
         $metric1->setName('sum_time_server');
         $metricsList->addMetric($metric1);