--- a/plugins/PagePerformance/Columns/TimeTransfer.php
+++ b/plugins/PagePerformance/Columns/TimeTransfer.php
@@ -30,7 +30,7 @@ class TimeTransfer extends Base
 
     public function configureMetrics(MetricsList $metricsList, DimensionMetricFactory $dimensionMetricFactory)
     {
-        $metric1 = $dimensionMetricFactory->createMetric(ArchivedMetric::AGGREGATION_SUM);
+        $metric1 = $dimensionMetricFactory->createMetric('sum(' . $this->getSqlCappedValue() . ')');
         $metric1->setName('sum_time_transfer');
         $metricsList->addMetric($metric1);