--- a/plugins/PagePerformance/Metrics.php
+++ b/plugins/PagePerformance/Metrics.php
@@ -9,7 +9,7 @@
 namespace Piwik\Plugins\PagePerformance;
 
 use Piwik\Columns\Dimension;
-use Piwik\Plugin\Dimension\ActionDimension;
+use Piwik\Plugins\PagePerformance\Columns\Base;
 use Piwik\Plugins\PagePerformance\Columns\Metrics\AveragePageLoadTime;
 use Piwik\Plugins\PagePerformance\Columns\Metrics\AverageTimeDomCompletion;
 use Piwik\Plugins\PagePerformance\Columns\Metrics\AverageTimeDomProcessing;
@@ -97,7 +97,7 @@ class Metrics
         $table = 'log_link_visit_action';
 
         /**
-         * @var ActionDimension[] $performanceDimensions
+         * @var Base[] $performanceDimensions
          */
         $performanceDimensions = [
             new TimeNetwork(),
@@ -112,12 +112,7 @@ class Metrics
             $column = $table . '.' . $id;
             $metricsConfig['sum_' . $id] = [
                 'aggregation' => 'sum',
-                'query' => "sum(
-                    case when " . $column . " is null
-                        then 0
-                        else " . $column . "
-                    end
-                ) / 1000"
+                'query' => "sum(" . sprintf($dimension->getSqlCappedValue(), $column) . ") / 1000"
             ];
             $metricsConfig['nb_hits_with_' . $id] = [
                 'aggregation' => 'sum',