@sgiehl 2023-05-15
Hi @rr-it,
Thanks for your contribution. I had a look across the code changes and I'm not totally sure what they are meant for.
To confirm I've understood everything correct:

This PR introduces some sort of maximal values for the performance metrics that are used in calculations.
So if there were values tracked higher than the one configured, the calculations will use the configured values instead.

Is that correct?