@rr-it 2023-05-15
@sgiehl That's perfectly correct.

Please also see my latest comment on #17847:
https://github.com/matomo-org/matomo/issues/17847#issuecomment-1546899519