@rr-it 2023-05-15
Devices, especially iPhones, sometimes send insane high performance values which don't make any sense.
E.g. DOM processing time of 4 hours instead of 250 microseconds.

Without this PR:
+ One value of 4 h and 10000 values of 250 ms result in an average of ~ `1.690 seconds`
  + $`4h = 4×60×60×1000 ms = 14400000 ms`$
  + $`\frac{14400000 ms + 10000×250 ms}{10001} = 1690 ms`$
+ 10000 values of 250 ms result in an average of `0.250 seconds`
     $`\frac{10000×250 ms}{10000} = 250 ms`$

With this PR:
`time_dom_processing_cap_duration_ms = 50000`

+ One value of 4 h and 10000 values of 250 ms result in an average of ~ `0.255 seconds`
  + $`4h = 4×60×60×1000 ms = 14400000 ms`$
    $`14400000 ms`$ capped to $`50000 ms`$
  + $`\frac{50000 ms + 10000×250 ms}{10001} = 255 ms`$
+ 10000 values of 250 ms result in an average of `0.250 seconds`
     $`\frac{10000×250 ms}{10000} = 250 ms`$