@Stan-vw 2023-06-26
I think we would like this report to not show the outliers (top & bottom 1% sounds like a fine test to begin with, but as you say that could be more implementation work maybe we can pick fixed values from historic reports).
However, I think we should still have a report that shows all values (if we don't have that already). I've attached some potential ways to visualise this, albeit that I didn't add the relevant axes (days vs seconds):
![image](https://github.com/matomo-org/matomo/assets/52564377/72547714-8f81-43ba-aa20-c430967c8496)
![image](https://github.com/matomo-org/matomo/assets/52564377/d8c6ec35-c5f6-4249-a04b-eb2df780247f)