@michalkleiner 2023-07-04
The business logic still needs to be resolved (whether we do fixed values, top and bottom percentage or something else). There's also a merge conflict now which needs to be resolved, plus a change request from Stefan.

I've removed the Needs review tag for now until all the above is addressed at which point it can be marked as Needs review again.