@rr-it 2023-07-04
On the business logic - most points were already mentioned:

+ Clear preference to cap by top percentage over fixed values:
  + Top percentage is more implementation work.
  + Fixed values already archive the target to make the chart meaningful. (Right now the charts informational value is insignificant.)
+ Top and bottom? No need for bottom value caps as the impact of low values is neglectable.

## Without this PR

### No outliers

+ 10000 values of 250 ms result in an average of `0.250 seconds`
     $`\frac{10000×250 ms}{10000} = 250 ms`$

### Top outlier

+ One value of 4 h and 10000 values of 250 ms result in an average of ~ `1.690 seconds`
  + $`4h = 4×60×60×1000 ms = 14400000 ms`$
  + $`\frac{1 ms + 10000×250 ms}{10001} = 1690 ms`$

### Bottom outlier

+ One value of 1 ms and 10000 values of 250 ms result in an average of ~ `0.250 seconds`
  + $`\frac{14400000 ms + 10000×250 ms}{10001} = 249.975 ms`$