@Stan-vw 2023-07-04
I understand the implication of the math and agree that the reports' value/reliability diminishes drastically if outliers are included in the calculation. My challenges with giving a clear conclusive answer are:
1. I am not yet familiar with the report we're talking about, and the different ways users might be leveraging it. I'm wondering what type of performance are people tracking: Website pageloads are in milliseconds, but what if people are tracking items similar to our archiving processes which can take much longer? 
2. If we are setting the cut-off values, they ultimately are arbitrary for users. Can we make the lower/upper values a variable they can set themselves? (either per report or as a whole)
3. One of our pillars is to show all data so I'm hesitant to simply greenlight a change that hides some data, especially when we're not telling users about it. Can we add something like "X1 data points were hidden as they are below the outlier threshold of Y1 milliseconds, X2 data points were hidden as they are above the outlier threshold of Y2 milliseconds"?
4. Given that people might have an edge case where outliers on either side are of interest, can we give users the ability to turn this off in a config?
5. Would it be a decent idea to make this a separate report, so that people can put "Performance report" and "Performance report (excluding outliers)" next to each other?

Thanks in advance for helping me understand the feature and how people are using it more deeply.