@Stan-vw 2023-07-04
Thanks @michalkleiner for pointing me to the [area of this PR that answers all my questions](https://github.com/matomo-org/matomo/pull/20718/files#diff-72968590ea965bba4c1ecf90875001330282aeaa9278c468336a96a6d683ab0e)

If I understand correctly, the current PR already allows people to:
1. Choose their own values
2. Turn this feature on/off

Knowing this, I am happy for this to be added as an opt-in feature (from @sgiehl 's message above I interpret it's gone through review and testing).

In the current state (as far as I understand it), I would not ship this as a default state because people aren't aware it's hiding data. I think it could be appropriate as a default state when we can clearly indicate how many datapoints were excluded from the calculation of a metric in some user friendly way (some non-vetted ideas that comes to mind: show "avg load time of 250 ms, excluding 3 outliers above 10000 ms" on hover-over or on a footnote).