@michalkleiner 2023-07-04
Thanks @Stan-vw, that's great. I also think making this an opt-in measure addresses all the concerns here and we can always iterate on that in future, while on-premise users will be able to configure this locally as they please.

@sgiehl what is the best way to provide the example config but not have it active, would it still be in the `global.ini.php` file but just commented out? Then the comment about disabling it via providing 0 values can be removed and the code adjusted to treat missing config as this being disabled.