@Stan-vw 2023-07-05
This might be interesting to add in the general settings? Have added a quick & dirty mockup of what this could look like. Not sure if that'd work though.
![image](https://github.com/matomo-org/matomo/assets/52564377/c80a4a0f-9693-430b-8bbd-809c2026d88c)