@bx80 2023-07-10
iPhone outliers could be related to different behaviour with the DOMContentLoaded event on IOS https://developer.apple.com/forums/thread/651215