@bx80 2023-07-11
>All performance metrics are just read from [PerformanceTiming API](https://developer.mozilla.org/en-US/docs/Web/API/PerformanceTiming) e.g. [loadEventEnd](https://developer.mozilla.org/en-US/docs/Web/API/PerformanceTiming/loadEventEnd)

Thanks @rr-it, I was curious about the cause of the high DOM loading times, but I now see that issue https://github.com/matomo-org/matomo/issues/17646 you linked in the related other ticket gives some idea of how these outlying values can originate :+1: