@michalkleiner 2023-08-14
Hi @rr-it,
thanks again for providing the PR, we do appreciate it. At the moment we are thinking whether the configuration should be done via a config or, as @Stan-vw mocked above, through the settings, so that people who don't have access to modifying the config files were able to also adjust the values.
Would that we an option that you'd be willing to look into as well?

We would also need to have the mechanism covered by a test to know that the caps are working correctly for future maintainability, are you familiar with how to provide one?